Move CLI into new file and add monitor command
